#675 refactor: extract performPlayerLeave into shared lib/lobby-leave.ts
